- Añadida la estructura GeometriaObj para definir características geométricas del objeto: posición del centroide, inercias, ángulo de rotación de ejes principales, factor de forma y k ideal. - En ImgDatos: Añadida nueva forma para llamar a a ImgDatos, creando sólo conect y contorno. Añadido el contorno exterior a un vector posPixcontornoext. Cálculo del perímetro. Cálculo de momentos y elipse en la función añadida CalculaMomentos. Añadida función Perfilamiento para reducir patas y antenas.

+++ TFM/Eje.h
@@ -19,10 +19,10 @@ public:
 public:
 	
 	/// DIBUJA EJE PRINCIPAL DE INERCIA Y LO DEVUELVE COMO IMAGEN
-	void DibujaEje(Eje& eje, cv::Vec2f& centroide, double angulo);
+	void DibujaEje(Eje& eje, cv::Vec2d& centroide, double angulo);
 
 	/// DIBUJA EJE SECUNDARIO
-	void DibujaEjeSec(Eje& eje, cv::Vec2f& centroide, double angulo);
+	void DibujaEjeSec(Eje& eje, cv::Vec2d& centroide, double angulo);
 
 	/// INTERSECCIÓN EJES - CONTORNO
 	double Intersecc(Eje& eje, cv::Mat& contorno); // Intersección eje longitudinal
